Moved out code from some mod.rs files to their own modules
It was getting annoying to understand what this files actually were in the code editor.
