+++ src/session/content/mod.rs
@@ -1,130 +1,5 @@
 mod chat_history;
+mod content;
 
 use self::chat_history::ChatHistory;
-
-use crate::session::Chat;
-use gtk::prelude::*;
-use gtk::subclass::prelude::*;
-use gtk::glib;
-
-mod imp {
-    use super::*;
-    use adw::subclass::prelude::BinImpl;
-    use gtk::CompositeTemplate;
-    use once_cell::sync::Lazy;
-    use std::cell::{Cell, RefCell};
-
-    #[derive(Debug, Default, CompositeTemplate)]
-    #[template(resource = "/com/github/melix99/telegrand/ui/content.ui")]
-    pub struct Content {
-        pub compact: Cell<bool>,
-        pub chat: RefCell<Option<Chat>>,
-    }
-
-    #[glib::object_subclass]
-    impl ObjectSubclass for Content {
-        const NAME: &'static str = "Content";
-        type Type = super::Content;
-        type ParentType = adw::Bin;
-
-        fn class_init(klass: &mut Self::Class) {
-            ChatHistory::static_type();
-            Self::bind_template(klass);
-
-            klass.install_action("content.go-back", None, move |widget, _, _| {
-                widget.set_chat(None);
-            });
-        }
-
-        fn instance_init(obj: &glib::subclass::InitializingObject<Self>) {
-            obj.init_template();
-        }
-    }
-
-    impl ObjectImpl for Content {
-        fn properties() -> &'static [glib::ParamSpec] {
-            static PROPERTIES: Lazy<Vec<glib::ParamSpec>> = Lazy::new(|| {
-                vec![
-                    glib::ParamSpec::new_boolean(
-                        "compact",
-                        "Compact",
-                        "Wheter a compact view is used or not",
-                        false,
-                        glib::ParamFlags::READWRITE,
-                    ),
-                    glib::ParamSpec::new_object(
-                        "chat",
-                        "Chat",
-                        "The chat currently shown",
-                        Chat::static_type(),
-                        glib::ParamFlags::READWRITE | glib::ParamFlags::EXPLICIT_NOTIFY,
-                    ),
-                ]
-            });
-
-            PROPERTIES.as_ref()
-        }
-
-        fn set_property(
-            &self,
-            obj: &Self::Type,
-            _id: usize,
-            value: &glib::Value,
-            pspec: &glib::ParamSpec,
-        ) {
-            match pspec.name() {
-                "compact" => {
-                    let compact = value.get().unwrap();
-                    self.compact.set(compact);
-                }
-                "chat" => {
-                    let chat = value.get().unwrap();
-                    obj.set_chat(chat);
-                }
-                _ => unimplemented!(),
-            }
-        }
-
-        fn property(&self, obj: &Self::Type, _id: usize, pspec: &glib::ParamSpec) -> glib::Value {
-            match pspec.name() {
-                "compact" => self.compact.get().to_value(),
-                "chat" => obj.chat().to_value(),
-                _ => unimplemented!(),
-            }
-        }
-
-        fn constructed(&self, obj: &Self::Type) {
-            self.parent_constructed(obj);
-        }
-    }
-
-    impl WidgetImpl for Content {}
-    impl BinImpl for Content {}
-}
-
-glib::wrapper! {
-    pub struct Content(ObjectSubclass<imp::Content>)
-        @extends gtk::Widget, adw::Bin;
-}
-
-impl Content {
-    pub fn new() -> Self {
-        glib::Object::new(&[]).expect("Failed to create Content")
-    }
-
-    pub fn chat(&self) -> Option<Chat> {
-        let priv_ = imp::Content::from_instance(self);
-        priv_.chat.borrow().clone()
-    }
-
-    fn set_chat(&self, chat: Option<Chat>) {
-        if self.chat() == chat {
-            return;
-        }
-
-        let priv_ = imp::Content::from_instance(self);
-        priv_.chat.replace(chat);
-
-        self.notify("chat");
-    }
-}
+pub use self::content::Content;
